♻️ :: floating notice move to design system
