+++ design-system/src/main/res/drawable/coloricnotice.xml
